Add files via upload
Updated to allow the (optional) addition of feed info to the request info. These are in the format of a JSON Dict 
E.g. {"longitude":90.9, "latitude":91.24, "tagNames":"car park|spaces", "unitText":"No of free spaces"}
For backward compatibility, defaults to {} where no feed info is provided.
This info is not required for calls to data hubs, but for example are used when populating a database such as InfluxDb.